+++ app/src/main/AndroidManifest.xml
@@ -58,24 +58,6 @@
                 android:name="android.support.PARENT_ACTIVITY"
                 android:value="de.projektss17.bonpix.A_Main" />
         </activity>
-        <activity
-            android:name=".A_Einstellungen"
-            android:label="@string/title_einstellungen_screen"
-            android:parentActivityName=".A_Main"
-            android:theme="@style/AppTheme.NoActionBar">
-            <meta-data
-                android:name="android.support.PARENT_ACTIVITY"
-                android:value="de.projektss17.bonpix.A_Main" />
-        </activity>
-        <activity
-            android:name=".A_Version"
-            android:label="@string/title_version_screen"
-            android:parentActivityName=".A_Einstellungen"
-            android:theme="@style/AppTheme.NoActionBar">
-            <meta-data
-                android:name="android.support.PARENT_ACTIVITY"
-                android:value="de.projektss17.bonpix.A_Einstellungen" />
-        </activity>
 
         <meta-data
             android:name="com.google.android.gms.vision.DEPENDENCIES"
@@ -91,7 +73,14 @@
                 android:name="android.support.PARENT_ACTIVITY"
                 android:value="de.projektss17.bonpix.A_Main" />
         </activity>
-        <activity android:name=".A_Einstellungen_test"></activity>
+        <activity android:name=".A_Einstellungen"
+            android:label="@string/title_einstellungen_screen"
+            android:parentActivityName=".A_Main"
+            android:screenOrientation="portrait">
+            <meta-data
+                android:name="android.support.PARENT_ACTIVITY"
+                android:value="de.projektss17.bonpix.A_Main" />
+        </activity>
     </application>
 
 </manifest>
