--- app/src/main/java/de/projektss17/bonpix/A_Einstellungen_test.java
@@ -1,64 +1,0 @@
-package de.projektss17.bonpix;
-
-import android.os.Build;
-import android.preference.PreferenceActivity;
-import android.os.Bundle;
-import android.support.v7.widget.Toolbar;
-import android.util.TypedValue;
-import android.view.LayoutInflater;
-import android.view.View;
-import android.view.ViewGroup;
-import android.widget.LinearLayout;
-import android.widget.ListView;
-
-
-public class A_Einstellungen_test extends PreferenceActivity {
-
-    @Override
-    protected void onCreate(Bundle savedInstanceState) {
-        super.onCreate(savedInstanceState);
-        setContentView(R.layout.box_einstellungen_test);
-        getFragmentManager().beginTransaction()
-                .replace(android.R.id.content, new F_Einstellungen())
-                .commit();
-    }
-
-    @Override
-    protected void onPostCreate(Bundle savedInstanceState) {
-
-        super.onPostCreate(savedInstanceState);
-        Toolbar bar;
-
-        if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.JELLY_BEAN) {
-            LinearLayout root = (LinearLayout) findViewById(android.R.id.list).getParent().getParent().getParent();
-            bar = (Toolbar) LayoutInflater.from(this).inflate(R.layout.box_einstellungen_toolbar, root, false);
-            root.addView(bar, 0);
-        } else {
-            ViewGroup root = (ViewGroup) findViewById(android.R.id.content);
-            ListView content = (ListView) root.getChildAt(0);
-
-            root.removeAllViews();
-
-            bar = (Toolbar) LayoutInflater.from(this).inflate(R.layout.box_einstellungen_toolbar, root, false);
-
-            int height;
-            TypedValue tv = new TypedValue();
-            if (getTheme().resolveAttribute(R.attr.actionBarSize, tv, true)) {
-                height = TypedValue.complexToDimensionPixelSize(tv.data, getResources().getDisplayMetrics());
-            }else{
-                height = bar.getHeight();
-            }
-
-            content.setPadding(0, height, 0, 0);
-            root.addView(content);
-            root.addView(bar);
-        }
-
-        bar.setNavigationOnClickListener(new View.OnClickListener() {
-            @Override
-            public void onClick(View v) {
-                finish();
-            }
-        });
-    }
-}
