+++ app/src/main/res/layout/box_einstellungen_content.xml
@@ -1,68 +1,14 @@
 <?xml version="1.0" encoding="utf-8"?>
-<RelativeLayout xmlns:android="http://schemas.android.com/apk/res/android"
+<android.support.design.widget.CoordinatorLayout
+    xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:app="http://schemas.android.com/apk/res-auto"
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    app:layout_behavior="@string/appbar_scrolling_view_behavior"
-    tools:context="de.projektss17.bonpix.A_Einstellungen"
-    tools:showIn="@layout/box_einstellungen_screen">
+    tools:context=".A_Einstellungen">
 
-    <View
-        android:layout_width="match_parent"
-        android:layout_height="2px"
-        android:layout_marginTop="45dp"
-        android:background="@color/linie"
-        android:visibility="visible" />
+    <ListView android:id="@android:id/list"
+        android:layout_height="0dp"
+        android:layout_width="0dp"/>
 
-
-    <ImageButton
-        android:id="@+id/einstellungen_version_button"
-        android:layout_width="wrap_content"
-        android:layout_height="@dimen/standard_30"
-        android:layout_alignParentEnd="true"
-        android:layout_alignParentLeft="true"
-        android:layout_alignParentRight="true"
-        android:layout_alignParentStart="true"
-        android:layout_marginLeft="10dp"
-        android:layout_marginTop="60dp"
-        android:background="?android:attr/colorBackground"
-        android:scaleType="fitStart"
-        app:srcCompat="@mipmap/ic_info_outline_black_24dp" />
-
-    <TextView
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:layout_alignBottom="@+id/einstellungen_version_button"
-        android:layout_alignParentTop="true"
-        android:layout_marginLeft="55dp"
-        android:layout_marginTop="61dp"
-        android:clickable="false"
-        android:gravity="start"
-        android:text="@string/menu_einstellungen_version"
-        android:textColor="@android:color/black"
-        android:textSize="20sp"
-        android:id="@+id/textView">
-    </TextView>
-
-    <View
-        android:layout_width="match_parent"
-        android:layout_height="2px"
-        android:layout_marginTop="100dp"
-        android:layout_marginLeft="50dp"
-        android:background="@color/linie"
-        android:visibility="visible" />
-
-    <Switch
-        android:id="@+id/switch1"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_alignLeft="@+id/textView"
-        android:layout_alignStart="@+id/textView"
-        android:layout_centerVertical="true"
-        android:layout_marginLeft="85dp"
-        android:layout_marginStart="85dp"
-        android:text="TestSwitch " />
-
-
-</RelativeLayout>
+</android.support.design.widget.CoordinatorLayout>
