--- app/src/main/res/layout/box_einstellungen_test.xml
@@ -1,14 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<android.support.design.widget.CoordinatorLayout
-    xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
-    xmlns:tools="http://schemas.android.com/tools"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    tools:context=".A_Einstellungen_test">
-
-    <ListView android:id="@android:id/list"
-        android:layout_height="0dp"
-        android:layout_width="0dp"/>
-
-</android.support.design.widget.CoordinatorLayout>
