--- app/src/main/res/layout/box_version_content.xml
@@ -1,34 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<RelativeLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
-    xmlns:tools="http://schemas.android.com/tools"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    app:layout_behavior="@string/appbar_scrolling_view_behavior"
-    tools:context="de.projektss17.bonpix.A_Version"
-    tools:showIn="@layout/box_version_screen">
-
-    <TextView
-        android:id="@+id/Versionakt"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_alignParentTop="true"
-        android:layout_centerHorizontal="true"
-        android:layout_marginTop="125dp"
-        android:text="@string/Versionakt"
-        android:textSize="30sp"
-        tools:layout_editor_absoluteX="142dp"
-        tools:layout_editor_absoluteY="180dp" />
-
-    <TextView
-        android:id="@+id/Versionnr"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:text="@string/built_nr"
-        android:textSize="30sp"
-        tools:layout_editor_absoluteX="178dp"
-        tools:layout_editor_absoluteY="283dp"
-        android:layout_centerVertical="true"
-        android:layout_centerHorizontal="true" />
-
-</RelativeLayout>
